AI CV-Evaluator/app/requirements.txt
fastapi
uvicorn
sqlmodel
passlib[bcrypt]
openai
python-dotenv
streamlit
requests